+++ src/main/java/com/learning/dao/MessageRepository.java
@@ -1,6 +1,6 @@
-package org.ac.cst8277.fiskin.mehmet.dao;
+package com.learning.dao;
 
-import org.ac.cst8277.fiskin.mehmet.entity.MessageEntity;
+import com.learning.entity.MessageEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
