+++ src/main/java/com/learning/dao/RoleRepository.java
@@ -1,6 +1,6 @@
-package org.ac.cst8277.fiskin.mehmet.dao;
+package com.learning.dao;
 
-import org.ac.cst8277.fiskin.mehmet.entity.RoleEntity;
+import com.learning.entity.RoleEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
