+++ src/test/java/com/learning/FinalUserManagementSystemApplicationTests.java
@@ -1,4 +1,4 @@
-package org.ac.cst8277.fiskin.mehmet;
+package com.learning;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
